calendar sql db

+++ calendar/app/domain/model/event.go
@@ -4,12 +4,12 @@ import "time"
 
 //Event Calendar Event Model
 type Event struct {
-	ID          uint32    `json:"id"`
+	ID          int64     `json:"id"`
 	Title       string    `json:"title"`
 	Description string    `json:"description"`
 	Location    string    `json:"location"`
 	StartTime   time.Time `json:"startTime"`
 	EndTime     time.Time `json:"endTime"`
-	UserID      uint32    `json:"userId"`
-	CalendarID  uint32    `json:"calendarId"`
+	UserID      int64     `json:"userId"`
+	CalendarID  int64     `json:"calendarId"`
 }
